rt/server: full 64-bit random seeds (#358)
The PRNG has been mt19937_64 all along, but the captured seed was one
32-bit random_device draw, capping the space of reproducible random
sequences at 2^32. Widen it to 64 bits end to end: the context's
OptRandomSeed (seeded from two random_device draws), TLazyRandomList's
ctors (mt19937_64's native seed width, so the seed reaches the PRNG
unnarrowed), the TMetaRecord::TEntry capture, and the Tetris
assertion-replay reconstruction.

The issue's other half -- rolling the captured Now() into the assertion
capture -- turned out to already exist: the commit path stores
GetOptNow() and the seed in every update's meta record, and
TestAssertions reconstructs the replay context from
GetRunTimestamp()/GetRandomSeed(). The stale 2014 TODO is replaced with
a statement of that fact.

Widening the meta record's seed field changes its serialized shape;
databases written by older builds would not read the entry back as this
type (no cross-build on-disk compatibility is promised in this tree).

Closes #358.

+++ orly/server/meta_record.cc
@@ -31,7 +31,7 @@ RECORD_ELEM(TMetaRecord::TEntry, string, MethodName);
 RECORD_ELEM(TMetaRecord::TEntry, TMetaRecord::TEntry::TArgByName, ArgByName);
 RECORD_ELEM(TMetaRecord::TEntry, TMetaRecord::TEntry::TExpectedPredicateResults, ExpectedPredicateResults);
 RECORD_ELEM(TMetaRecord::TEntry, Base::Chrono::TTimePnt, RunTimestamp);
-RECORD_ELEM(TMetaRecord::TEntry, uint32_t, RandomSeed);
+RECORD_ELEM(TMetaRecord::TEntry, uint64_t, RandomSeed);
 
 /* Metadata for TMetaRecord. */
 RECORD_ELEM(TMetaRecord, TMetaRecord::TEntryByUpdateId, EntryByUpdateId);
